+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## [0.10.1] - 2017-08-15
+
+### Fixed
+- [#88] Fixed fatal error caused when resolving request objects out of the service container when there was no 
+inbound request bound in the service container.
+
 ## [0.10.0] - 2017-07-29
 
 ### Added